Add explicit parser for iCalendar and refactor exception handling
